requirements.txt
Flask==2.2.5
pyserial==3.5
